moved waste stuff out of waste comp folder
