+++ yoso/client/src/components/WasteItem.js
